# apps/server/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: SiteCRM API
  version: 0.2.0
  description: |
    Primary REST API for SiteCRM.  Handles authentication, lead management,
    outreach, notifications, site generation, and deployments.

    **Authentication:** Most endpoints require a Bearer JWT.  Obtain a token
    via `POST /api/auth/login` or `POST /api/auth/register` and pass it as
    `Authorization: Bearer <token>` on subsequent requests.

    **Common error responses (apply to every operation unless noted):**
    - `400 Bad Request` — request body or query failed validation
      (`additionalProperties: false` rejects unknown fields).
    - `401 Unauthorized` — missing, invalid, or expired token on a protected route.
    - `503 Service Unavailable` — the database is not configured/reachable.

    All errors use the shared `Error` schema. Reusable definitions are under
    `components/responses` (`BadRequest`, `Unauthorized`, `NotFound`,
    `ServiceUnavailable`).
  contact:
    name: Riverfront Labs
servers:
  - url: http://api.localhost
    description: Local development (Traefik)
  - url: http://localhost:3000
    description: Direct local development
...
paths:
...
  /api/leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads with optional filters
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
        - name: status
          in: query
          description: Filter by one or more statuses (repeat the param for multiple).
          schema:
            oneOf:
              - $ref: '#/components/schemas/LeadStatus'
              - type: array
                items:
                  $ref: '#/components/schemas/LeadStatus'
        - name: shortlisted
          in: query
          schema: { type: boolean }
        - name: search
          in: query
          description: Case-insensitive substring match on name, email, or location.
          schema: { type: string, maxLength: 200 }
      responses:
        '200':
          description: Paginated lead list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedLeads'

    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewLead'
      responses:
        '201':
          description: Lead created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
  /api/leads/{leadId}:
    parameters:
      - name: leadId
        in: path
        required: true
        schema: { type: string, format: uuid }

    get:
      operationId: getLead
      tags: [leads]
      summary: Get a single lead
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lead found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '404':
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update any subset of mutable lead fields
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLead'
      responses:
        '200':
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '404':
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead (cascades to notes and contact events)
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
        '404':
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/leads/{leadId}/notes:
    parameters:
      - name: leadId
        in: path
        required: true
        schema: { type: string, format: uuid }

    get:
      operationId: listNotes
      tags: [leads]
      summary: List notes for a lead (newest first)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Note'
        '404':
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      operationId: createNote
      tags: [leads]
      summary: Append a note to a lead
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 10000
      responses:
        '201':
          description: Note created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '404':
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/leads/{leadId}/notes/{noteId}:
    parameters:
      - name: leadId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: noteId
        in: path
        required: true
        schema: { type: string, format: uuid }

    delete:
      operationId: deleteNote
      tags: [leads]
      summary: Delete a specific note
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
        '404':
          description: Lead or note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/leads/{leadId}/events:
    parameters:
      - name: leadId
        in: path
        required: true
        schema: { type: string, format: uuid }

    get:
      operationId: listContactEvents
      tags: [leads]
      summary: List contact events for a lead (newest first)
      description: |
        Immutable audit log of every outreach touch-point — emails sent,
        DMs sent/received, calls, meetings. Events are created by the outreach
        system (Resend webhooks, Meta Graph API webhooks) or manually.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Events list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContactEvent'
        '404':
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Notifications ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # Reusable error responses — reference from operations via
  # `$ref: '#/components/responses/<Name>'`. See the info description for when
  # each applies across the API.
...
  schemas:
...
    LeadStatus:
      type: string
      enum: [new, contacted, qualified, proposal, won, lost]
    OutreachChannel:
      type: string
      enum: [email, facebook_dm, instagram_dm]
      description: SMS is intentionally excluded (regulatory constraints).
    ContactEventType:
      type: string
      enum:
        - email_sent
        - email_opened
        - email_clicked
        - facebook_dm_sent
        - facebook_dm_received
        - instagram_dm_sent
        - instagram_dm_received
        - call
        - meeting
    Lead:
      type: object
      required: [id, userId, name, status, shortlisted, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: ['string', 'null']
          format: email
        website:
          type: ['string', 'null']
        facebookPageId:
          type: ['string', 'null']
          description: Facebook Page ID for Messenger outreach.
        instagramAccountId:
          type: ['string', 'null']
          description: Instagram Business Account ID for DM outreach.
        websiteScore:
          type: ['number', 'null']
          description: 0–100 score from the intelligence website analyser.
        websiteGrade:
          type: ['string', 'null']
        websiteNotes:
          type: ['string', 'null']
        score:
          type: ['number', 'null']
          description: Composite intelligence score (0–100).
        rating:
          type: ['number', 'null']
          description: Google Places star rating (1–5).
        reviews:
          type: ['integer', 'null']
        type:
          type: ['string', 'null']
          description: Business category (e.g. "restaurant", "law firm").
        location:
          type: ['string', 'null']
        placeId:
          type: ['string', 'null']
          description: Google Places ID.
        linkedSiteId:
          type: ['string', 'null']
          format: uuid
        status:
          $ref: '#/components/schemas/LeadStatus'
        shortlisted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    NewLead:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        email:
          type: string
          format: email
          maxLength: 255
        website:
          type: string
          maxLength: 512
        facebookPageId:
          type: string
          maxLength: 255
        instagramAccountId:
          type: string
          maxLength: 255
        type:
          type: string
          maxLength: 100
        location:
          type: string
          maxLength: 255
        placeId:
          type: string
          maxLength: 255
        status:
          $ref: '#/components/schemas/LeadStatus'
        shortlisted:
          type: boolean
    UpdateLead:
      type: object
      additionalProperties: false
      minProperties: 1
      description: All fields are optional — only supplied fields are updated.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        email:
          type: ['string', 'null']
          format: email
          maxLength: 255
        website:
          type: ['string', 'null']
          maxLength: 512
        facebookPageId:
          type: ['string', 'null']
          maxLength: 255
        instagramAccountId:
          type: ['string', 'null']
          maxLength: 255
        websiteScore:
          type: ['number', 'null']
          minimum: 0
          maximum: 100
        websiteGrade:
          type: ['string', 'null']
          maxLength: 10
        websiteNotes:
          type: ['string', 'null']
          maxLength: 1000
        score:
          type: ['number', 'null']
          minimum: 0
          maximum: 100
        rating:
          type: ['number', 'null']
          minimum: 0
          maximum: 5
        reviews:
          type: ['integer', 'null']
          minimum: 0
        type:
          type: ['string', 'null']
          maxLength: 100
        location:
          type: ['string', 'null']
          maxLength: 255
        placeId:
          type: ['string', 'null']
          maxLength: 255
        status:
          $ref: '#/components/schemas/LeadStatus'
        shortlisted:
          type: boolean
        linkedSiteId:
          type: ['string', 'null']
          format: uuid
    PaginatedLeads:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Lead'
        total:
          type: integer
          description: Total matching leads across all pages.
        page:
          type: integer
        pageSize:
          type: integer
    Note:
      type: object
      required: [id, leadId, content, createdAt]
      properties:
        id:
          type: string
          format: uuid
        leadId:
          type: string
          format: uuid
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    ContactEvent:
      type: object
      required: [id, leadId, type, channel, direction, createdAt]
      properties:
        id:
          type: string
          format: uuid
        leadId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/ContactEventType'
        channel:
          $ref: '#/components/schemas/OutreachChannel'
        direction:
          type: string
          enum: [sent, received]
        detail:
          type: ['string', 'null']
        externalId:
          type: ['string', 'null']
          description: Provider message ID used for deduplication.
        createdAt:
          type: string
          format: date-time

    # ── Notifications ─────────────────────────────────────────────────────────
...
    Error:
      type: object
      required: [statusCode, error, message]
      description: Standard error envelope — matches Fastify's default error shape.
      properties:
        statusCode:
          type: integer
        error:
          type: string
          description: HTTP reason phrase, e.g. `Not Found`
        message:
          type: string
          description: Human-readable explanation of what went wrong